Abstracted generic functionality to a parent class for services
