+++ solana/program_v2/programs/gateway_v2/src/instructions/admin/close_network.rs
